Added a Display * argument to make_req.

+++ lib/guidrawpoint.c
@@ -1,4 +1,4 @@
-/* $Id: guidrawpoint.c,v 1.4 2020/02/29 02:54:05 rkiesling Exp $ -*-c-*-*/
+/* $Id: guidrawpoint.c,v 1.5 2020/02/29 10:21:16 rkiesling Exp $ -*-c-*-*/
 
 /*
   This file is part of Ctalk.
@@ -85,7 +85,7 @@ int __ctalkX11PaneDrawPointBasic (void *d, int drawable_id,
 	   ":", pen_color,
 	   NULL);
   
-  make_req (shm_mem, (uintptr_t)d, PANE_DRAW_POINT_REQUEST,
+  make_req (shm_mem, d, PANE_DRAW_POINT_REQUEST,
    	    drawable_id, gc_ptr, d_buf);
 #ifdef GRAPHICS_WRITE_SEND_EVENT
   send_event.xgraphicsexpose.type = GraphicsExpose;
